Use new secgate API.
