+++ Cargo.toml
@@ -21,6 +21,7 @@ members = [
     "src/rt/monitor-api",
     "src/rt/monitor/tests/montest",
     "src/rt/monitor/tests/montest-lib",
+    "src/rt/monitor/tests/montest-srv",
     "src/rt",
     "src/rt/reference",
     "src/rt/minimal",
@@ -42,7 +43,6 @@ members = [
     "src/bin/cache",
     "src/srv/display-srv",
     "src/lib/twizzler-display",
-    "src/lib/twizzler-display/display-core",
 
     "src/ports/lwext4-rs",
     #"src/ports/rusqlite",
@@ -97,7 +97,7 @@ initrd = [
     "crate:montest",
     "lib:naming-srv",
     "lib:twz-rt",
-    "lib:montest-lib",
+    "lib:montest-srv",
     "lib:logboi-srv",
     "lib:pager-srv",
     "lib:cache-srv",
