+++ src/lib/logboi/Cargo.toml
@@ -10,4 +10,3 @@ crate-type = ["rlib"]
 secgate = { path = "../../lib/secgate" }
 twizzler-runtime = { path = "../../rt" }
 twizzler-rt-abi = "0.99"
-logboi-srv = { path = "../../srv/logboi-srv" }
