+++ src/lib/naming/src/lib.rs
@@ -1,6 +1,3 @@
-#[link(name = "naming_srv")]
-extern "C" {}
-
 use naming_core::{api::NamerAPI, handle::NamingHandle, Result};
 pub use naming_core::{dynamic::*, GetFlags, NsNode, NsNodeKind};
 use secgate::util::Descriptor;
@@ -8,45 +5,69 @@ use twizzler_rt_abi::object::ObjID;
 
 pub struct StaticNamingAPI {}
 
+#[secgate::gatecall]
+fn put(desc: Descriptor, name_len: usize, id: ObjID) -> Result<()> {}
+#[secgate::gatecall]
+fn get(desc: Descriptor, name_len: usize, flags: GetFlags) -> Result<NsNode> {}
+#[secgate::gatecall]
+fn open_handle() -> Result<(Descriptor, ObjID)> {}
+#[secgate::gatecall]
+fn close_handle(desc: Descriptor) -> Result<()> {}
+#[secgate::gatecall]
+fn enumerate_names(desc: Descriptor, name_len: usize) -> Result<usize> {}
+#[secgate::gatecall]
+fn enumerate_names_nsid(desc: Descriptor, id: ObjID) -> Result<usize> {}
+#[secgate::gatecall]
+fn remove(desc: Descriptor, name_len: usize) -> Result<()> {}
+#[secgate::gatecall]
+fn change_namespace(desc: Descriptor, name_len: usize) -> Result<()> {}
+#[secgate::gatecall]
+fn mkns(desc: Descriptor, name_len: usize, persist: bool) -> Result<()> {}
+#[secgate::gatecall]
+fn link(desc: Descriptor, name_len: usize, link_len: usize) -> Result<()> {}
+
+#[secgate::gatecall]
+pub fn namer_start(bootstrap: ObjID) -> Result<ObjID> {}
+
 impl NamerAPI for StaticNamingAPI {
     fn put(&self, desc: Descriptor, name_len: usize, id: ObjID) -> Result<()> {
-        naming_srv::put(desc, name_len, id)
+        put(desc, name_len, id)
     }
 
     fn get(&self, desc: Descriptor, name_len: usize, flags: GetFlags) -> Result<NsNode> {
-        naming_srv::get(desc, name_len, flags)
+        get(desc, name_len, flags)
     }
 
     fn open_handle(&self) -> Result<(Descriptor, ObjID)> {
-        naming_srv::open_handle()
+        open_handle()
     }
 
     fn close_handle(&self, desc: Descriptor) -> Result<()> {
-        naming_srv::close_handle(desc)
+        close_handle(desc)
     }
 
     fn enumerate_names(&self, desc: Descriptor, name_len: usize) -> Result<usize> {
-        naming_srv::enumerate_names(desc, name_len)
+        enumerate_names(desc, name_len)
     }
 
     fn enumerate_names_nsid(&self, desc: Descriptor, id: ObjID) -> Result<usize> {
-        naming_srv::enumerate_names_nsid(desc, id)
+        enumerate_names_nsid(desc, id)
     }
 
     fn remove(&self, desc: Descriptor, name_len: usize) -> Result<()> {
-        naming_srv::remove(desc, name_len)
+        remove(desc, name_len)
     }
 
     fn change_namespace(&self, desc: Descriptor, name_len: usize) -> Result<()> {
-        naming_srv::change_namespace(desc, name_len)
+        change_namespace(desc, name_len)
     }
 
     fn mkns(&self, desc: Descriptor, name_len: usize, persist: bool) -> Result<()> {
-        naming_srv::mkns(desc, name_len, persist)
+        mkns(desc, name_len, persist)
     }
 
     fn link(&self, desc: Descriptor, name_len: usize, link_len: usize) -> Result<()> {
-        naming_srv::link(desc, name_len, link_len)
+        link(desc, name_len, link_len)
     }
 }
