+++ src/lib/pager/src/lib.rs
@@ -1,12 +1,18 @@
+use secgate::{util::Descriptor, TwzError};
 use twizzler_rt_abi::object::ObjID;
 
-#[link(name = "pager_srv")]
-extern "C" {}
+#[secgate::gatecall]
+pub fn pager_start(q1: ObjID, q2: ObjID) -> Result<ObjID, TwzError> {}
 
-pub fn pager_start(q1: ObjID, q2: ObjID) {
-    pager_srv::pager_start(q1, q2).ok().unwrap();
-}
+#[secgate::gatecall]
+pub fn adv_lethe() -> Result<(), TwzError> {}
 
-pub fn adv_lethe() {
-    pager_srv::adv_lethe().unwrap();
-}
+#[secgate::gatecall]
+pub fn disk_len(id: ObjID) -> Result<u64, TwzError> {}
+
+#[secgate::gatecall]
+pub fn pager_open_handle() -> Result<(Descriptor, ObjID), TwzError> {}
+#[secgate::gatecall]
+pub fn pager_close_handle(desc: Descriptor) -> Result<(), TwzError> {}
+#[secgate::gatecall]
+pub fn pager_enumerate_external(desc: Descriptor, id: ObjID) -> Result<usize, TwzError> {}
