+++ src/lib/sgtest/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "sgtest"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+secgate = { path = "../../lib/secgate" }
