+++ src/rt/monitor/tests/montest-srv/Cargo.toml
@@ -1,12 +1,16 @@
 [package]
-name = "montest-lib"
+name = "montest-srv"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+crate-type = ["cdylib"]
 
 [dependencies]
 secgate = { path = "../../../../lib/secgate" }
 twizzler-runtime = { path = "../../.." }
 twizzler-rt-abi = { path = "../../../../abi/rt-abi/" }
 twizzler-abi = { path = "../../../../lib/twizzler-abi" }
+montest-lib = { path = "../montest-lib" }
