+++ src/rt/monitor/tests/montest/Cargo.toml
@@ -8,12 +8,12 @@ edition = "2021"
 [dependencies]
 twizzler-runtime = { path = "../../.." }
 secgate = { path = "../../../../lib/secgate" }
-montest-lib = { path = "../montest-lib" }
 monitor-api = { path = "../../../monitor-api" }
 ctor = "*"
 tracing = "0.1"
 tracing-subscriber = "0.3"
 twizzler-abi = { path = "../../../../lib/twizzler-abi" }
+montest-lib = { path = "../montest-lib" }
 
 [profile.release]
 debug = true
