+++ src/srv/devmgr-srv/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 # Important: this should be compiled as both an rlib, for exporting the trampoline,
 # and as a cdylib, as the actual .so file that will be exporting the gates.
 [lib]
-crate-type = ["cdylib", "rlib"]
+crate-type = ["cdylib"]
 
 [dependencies]
 twizzler-driver = { path = "../../lib/twizzler-driver" }
