+++ src/srv/display-srv/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 # Important: this should be compiled as both an rlib, for exporting the trampoline,
 # and as a cdylib, as the actual .so file that will be exporting the gates.
 [lib]
-crate-type = ["cdylib", "rlib"]
+crate-type = ["cdylib"]
 
 [dependencies]
 twizzler-abi = { path = "../../lib/twizzler-abi" }
@@ -17,4 +17,4 @@ twizzler-rt-abi = "0.99"
 tracing-subscriber = "*"
 tracing = "*"
 virtio-gpu = { path = "../../lib/virtio-gpu" }
-display-core = { path = "../../lib/twizzler-display/display-core" }
+twizzler-display = { path = "../../lib/twizzler-display" }
