+++ src/srv/logboi-srv/Cargo.toml
@@ -6,12 +6,13 @@ edition = "2021"
 # Important: this should be compiled as both an rlib, for exporting the trampoline,
 # and as a cdylib, as the actual .so file that will be exporting the gates.
 [lib]
-crate-type = ["cdylib", "rlib"]
+crate-type = ["cdylib"]
 
 [dependencies]
 secgate = { path = "../../lib/secgate" }
 twizzler-runtime = { path = "../../rt" }
 twizzler-abi = { path = "../../lib/twizzler-abi" }
 twizzler-rt-abi = "0.99"
+logboi = { path = "../../lib/logboi" }
 
 [features]
