+++ src/srv/pager-srv/src/lib.rs
@@ -41,8 +41,6 @@ mod request_handle;
 mod stats;
 mod threads;
 
-pub use handle::{pager_close_handle, pager_open_handle};
-
 /***
  * Tracing Init
  */
@@ -213,18 +211,18 @@ fn do_pager_start(q1: ObjID, q2: ObjID) -> ObjID {
     return bootstrap_id.into();
 }
 
-#[secgate::secure_gate]
+#[secgate::entry(lib = "pager")]
 pub fn pager_start(q1: ObjID, q2: ObjID) -> Result<ObjID> {
     Ok(do_pager_start(q1, q2))
 }
 
-#[secgate::secure_gate]
+#[secgate::entry(lib = "pager")]
 pub fn adv_lethe() -> Result<()> {
     run_async(PAGER_CTX.get().unwrap().paged_ostore(None)?.flush()).unwrap();
     Ok(())
 }
 
-#[secgate::secure_gate]
+#[secgate::entry(lib = "pager")]
 pub fn disk_len(id: ObjID) -> Result<u64> {
     run_async(PAGER_CTX.get().unwrap().paged_ostore(None)?.len(id.raw()))
         // TODO: err
